backend/requirements.txt
django==5.1.2
psycopg==3.1
psycopg2-binary==2.9.10
daphne==4.1.2
channels==4.2.0
channels_redis==4.2.1
pyjwt==2.9.0
